chore: check we are in terminal before creating one
I tried to run `node-launchpad` in a headless VM and it said it couldn't find a terminal. We don't
need to find a terminal if we are already running in one, so the check here is changed slightly.

We will probably want to reconfigure the elevated privilege to be a separate thing from configuring
a terminal, but on Windows this will be a bit more complex and will better be done as a separate
piece of work.