fix: check node registry exists before deleting it
The `remove_file` command returns an error if the file we are attempting to delete does not exist.
This meant on Windows, subsequent runs of the `reset` command could produce an error. It did not
occur on Linux or macOS because the registry file is briefly recreated on those platforms so that
it can be assigned special permissions; this does not apply on Windows.