feat: configure winsw in `node-launchpad`
The node launchpad will attempt to detect if `winsw.exe` is available, and if it isn't, it will
download the binary to the application's data directory, and use the `WINSW_PATH` variable to set
the location. This removes the need for separate steps to download the binary and make it available
somewhere on `Path`.

There's also a refactor in here: two functions, `get_config_dir` and `get_launchpad_data_dir_path`
are moved to the `config` module. It seems like they are both specifically related to configuration
and it makes sense for them to be in the `config` module rather than `utils`.