ci: prevent s3 files being overwritten
If a released binary has been uploaded to S3, it should be prevented from being overwritten. If it
is being overwritten, it could very well be suggestive of a mistake somewhere. If you really want to
intentionally overwrite the file for some reason, remove it from the S3 bucket manually, then run
the workflow again.