chore: package released binaries by architecture
Previously when we performed a release, we would do a Github Release for each binary in a particular
selection. Now, the approach is to have a single Github Release for all the binaries. The assets for
the single release will be a collection of zip files containing all nine binaries for each target
architecture.

The new release process stipulates that each package name should contain the version `YYYY.MM.X.Y`,
where `X` is the number of the current release cycle, and `Y` is a counter within the cycle, which
will be incremented for each RC build that is required.

To support these requirements, a `package-arch` target is added. It produces packages in the form,
e.g., `2024.07.1.1.autonomi.aarch64-unknown-linux-musl.zip`. The release cycle numbers are provided
from a new file at the root of the repository. Those numbers should be updated or reset in a release
commit. The target either reads the version number from a `PACKAGE_VERSION` environment variable or
it will build it. The environment variable will be used in the release process, because the package
version will be required in multiple places.

The new process retains the upload of individual binaries to S3, so those packaging and upload
targets are retained, but they are renamed to `package-bin` and `upload-packaged-bin-to-s3`, to
distinguish them from the architecture packages.

The `upload-github-release-assets` target is completely removed because it was uploading the
individual binaries to their respective Github Releases, which is no longer relevant. The release
workflow will be updated accordingly, but not in this commit.

For testing the new targets, this commit only updates the simpler 'build and package release
artifacts' workflow; a subsequent commit will update the full release workflow.