chore: implement new process in release workflow
The release workflow is changed to implement the new release process, as it's defined in our RFC.

The workflow will be triggered manually, but jobs will only run if the branch being used is either
`stable`, or it has an `rc` prefix. This is so we can have complete control over when to push the
release out, because it may need to be coordinated with other things, like community announcements.

We now define four jobs:
1. build
    - runs for either stable or RC release
    - builds binaries for the seven supported architectures
2. s3-release
    - runs for either stable or RC release
    - packages individual binaries and uploads them to S3
    - runs after build job
3. github-release
    - runs for either stable or RC release
    - packages binaries per-architecture and creates a Github Release with the packages as assets
    - on an RC release the Github Release with be a pre-release
    - runs after build job
4. publish-crates
    - only runs for a stable release
    - use `release-plz` to publish crates and produce crate tags
    - runs after s3-release job

After the `build` job completes, `s3-release` and `github-release` do not depend on each other, so
they will run in parallel. The `publish-crates` job will wait for the completion of `s3-release`
(the reason is explained in a comment in the workflow, which I won't repeat here).

Now that `release-plz` is only used for publishing crates and tagging, we don't need to configure it
on a per-crate basis, so the configuration file is now very minimal.