chore: script to bump versions for release candidate
Use this script at the beginning of a release candidate cycle to bump crates to their new versions.

Initially, the `release-plz` tool is used, detecting changes and bumping each crate accordingly,
which should cover breaking changes (if they have been correctly specified with the rules from the
conventional commits specification). For each crate not bumped by `release-plz`, a 'safety' bump
increments the `PATCH` version of that crate. In both cases, the `-rc.1` pre-release specifier is
applied to the versions used.

The reason for the safety bump is to guarantee unique binary versions for each release cycle. If we
didn't bump, we could end up with binaries with the same version number as the previous stable
release. This could be the case even though the dependencies of the binary had been updated, which
would cause it to have different content from the previously released version. It is not ideal to
bump crates with no real material changes, but doing it this way is easier to manage.

Also remove previous version bumping scripts we don't need any more.