chore: provide utils for release candidate post
A markdown file provides a template that can be used for a post on Discourse for discussing the
current release candidate. There is also a Python script which gets a list of the closed PRs between
now and the last release. Some of the output from this can be used in the template.

Also provide a script for getting the current versions of crates and binaries. This can be handy if
you lost the output from the bump script.