chore: remove s3 upload of 'latest' version
I don't think we are actively using these 'latest' versions any more. The only place that has
reference to it are some Ansible roles in `testnet-deploy`, but these also do not get actively used.
In the deployment process, we use specific versions. In the case where versions are not supplied,
the latest version is retrieved by querying `crates.io`.