chore: script for removing s3 bin archives
This is useful if something goes wrong with an RC or stable build or you just need to run the build
again and you want new binaries to be uploaded to S3.

Obviously it should be used with care, but in some situations it is necessary. It is extremely
tedious to remove the files manually, and often when trying to do that, one of them will be missed,
which can cause the whole run to fail.