chore: script to list merged pull requests
Give the script a file with a list of PR numbers and it will print various things that can be used
in the release candidate threads on Discourse.

Now that we're using merge commits, the list of PR numbers is easily obtainable from the comparison
of the `main` and `stable` branches, using `git log stable..main --oneline`.