fix: save node registry on every upgrade
Save the registry when each node is upgraded, whether there was an error or not. Previously, the
registry was only being saved after the entire suite was processed.

Waiting to the end was problematic during the deployment process when Ansible's SSH connection
was lost. It resulted in the registry not being updated and hence no nodes were being reported as
being upgraded. This led me on a bit of a wild goose chase when debugging the connection issues.

It will be very useful to have this fix in for the upgrade process, because nodes should then be
correctly reported as upgraded even if a connection is lost.