docs: provide more information on build-time keys
The documentation for build-time keys is updated to be readable by any developer. I think they were
previously written to be consumed more by internal developers, hence references to Github Secrets
and our release process.

I've re-arranged the section such that it describes how to build the client and node for the purpose
of connecting to the beta network. Information is also provided about the keys with respect to
building `safenode` from source; this is not just for developers, but any users who may want to
build from source.

I removed a section on releases: it was outdated and was more of a concern for internal developers.
We now have this documented in much more detail in our RFC for the release process.