fix: [#2032] `local status` cmd reports status correctly
In a service-based network, a node was determined to be running by searching for a process whose
path matched the path where the service binary is installed. This works because those are unique
paths. However, it did not work for a local network, because each process uses a binary at the same
path.

For a status refresh, we now distinguish between a service-based network and a local network. For
the local network, we determine if a node is running simply by connecting to its RPC service.