feat: shutting down on excess cpu does not return error
The node RPC is changed such that the stop event allows shutting the node down with either a success
or failure result.

Then, in the case of the excessive CPU usage, the node will be shut down with a success result. This
allows us to retain an on-failure restart policy for node-based services: with excess CPU usage,
they would shut down with a success and so would not be started again.