chore: include release cycle info in crate
For some reason the `sn_build_info` crate started to manifest an error that it couldn't find the
`release-cycle-info` file. The solution was to embed the information in the crate. This is less than
ideal, because it now needs to be updated in two places. However, we could have the other parts of
the release process read the Rust file. We can do that later.