fix: use `ant_networking` for metrics prefix
The `ant-networking` prefix was incompatible with the new version of Telegraf because it doesn't
allow hyphens in the metric name; however, in any case, this shouldn't have had the hyphen, and
should actually have remained as `sn_networking`.

It was mistakenly renamed, because we wanted to keep the `sn_` prefix in the metric names so that we
could change them more gradually. Now we've decided to just go with the `ant_` prefix, so we are
also changing the `sn_node` prefix to `ant_node` as part of this commit.