feat: provide new versioning mechanism
Implements the versioning scheme that was discussed and agreed.

It only uses the version from the `ant-protocol` crate; no keys or other external sources are
provided. Only the `MAJOR` and `MINOR` parts of the version are considered. The patch version can
advance without being considered.

We have discussed changing the version of other crates such that they are synchronised with this
protocol version.