chore: versioning for `ant` binary
This brings the `ant` binary in alignment with how versioning is handled for the other binaries.

As with the other binaries, to save a lot of hassle, the version arguments are implemented
'manually' rather than using the direct mechanism from `clap`. Therefore, to get the expected
behaviour, the main subcommand has to be made optional. From the user's point of view nothing
changes.

I took the opportunity to also sort the arguments alphabetically. This is how they are organised in
the node manager, and I find it makes for easier reference.