chore: switch back to github build agent
For the last several months we have been using a self-hosted build agent for the publishing step
because we encountered problems with disk space on the Github build agent. Now there is some kind of
issue with the connection from Github to the self-hosted agent.

Going to try switching back to Github infrastructure again. Perhaps we won't encounter the disk
space issue.